Change modified timestamps to match the old "changed" value
Started doing it manually, then realized it would take forever. Threw
together an extremely shitty python script using dateutil.parser and
made it output sed code, because yaml libraries don't know the concept
of in-place editing, they just rewrite the damn thing.

The script itself is too awful and non-reusable to publish, the only
part worth mentioning is the sed output, a bunch of lines like this:

/^  filename: "xmms.pl"/ {n; s/2008-05-17 17:39:09/2004-01-15 00:00:00/}

The "n" command grabs the next line. That was neat.

Scripts with modified date after 2011 weren't touched - most of the
incorrect dates are 2008 or 2009.

I reviewed the dateutil parsed dates - that library is really good.