Rename chanact.pl /[chanact_]window_(un)alias
The plugin adding /window_alias and /window_unalias commands creates
ambiguity when using the shorter /win (for /window) command thus the
need to spell out the whole /window command.

Since window aliases for chanact are seldom set and unset it is fine to
make /window_alias and /window_unalias tab-complete from the /chanact_
prefix. It also makes it easier to find those commands since the all of
the plugin's options live under /set chanact_ (as noted in the plugin's documentation itself.)