Fix network matching for query subcommand
In the absence of a network tag passed to the `query` subcommand, no
rules but the default would match and so the returned match(es) would
potentially not actually be the ones used in context of ctrlact's normal
operation.

This commit fixes #354 by special-casing the threshold match when the
data to be looked up doesn't actually specify a network tag. In this
case, the output for net will be '(ignored)'.

Signed-off-by: martin f. krafft <madduck@madduck.net>