Loosen version check on data file read
The versioning of the data file allows for making changes to the format,
but if the script version increases without the data file-format being
changed, things are expected to work as before, and thus there ought not
be an error spouted.

Let's only use the versioning to patch up handling for formats of the
past.

Signed-off-by: martin f. krafft <madduck@madduck.net>