trigger.pl: default address to empty in mask match
Server notices are included in the privnotices type, but their address
parameter is undefined, because the prefix of the message is just the
server name.  By defaulting address to an empty string in the masks
match, we can match server notices easily with `-privnotices -masks
'<servername>!'` instead of a really messy rawin + regexp.  (Irssi
merges the nick and address passed into masks_match with a `%s!%s`
format string, so there'll be a trailing "!" when address is an empty
string.)