autorejoin: Add settings, general cleanup
This merges the idea of autorejoin_channels from the version of the
script at http://wouter.coekaerts.be/irssi/scripts/autorejoin.pl

The version in this repository had different indentation and the timeout
delay stuff, so the end result is nothing like that one. Also that one
says Public Domain. Whatever.

This adds two settings:

- autorejoin_delay: time in seconds, defaults to 5, as the old code.
  Can be set to 0 to autorejoin instantly (like the other script)
- autorejoin_channels: channels whitelisted for autorejoin, applies to
  all channels if empty.

Not a fan of having this enabled for all channels, but it's the most
backwards-compatible behavior with the version of the script that was in
this repo. The other version defaulted to nothing.

This also changes the timeout_add into timeout_add_once and cleans up
parameter passing and acttag handling.

Also a couple of s/kick/kicked/ from the other script