Update quitrand.pl
Clean up the perl a little.  Remove an uneccesary loop, use a lexical file handle to avoid possible conflicts, and use chomp instead of chop.  chomp will only remove the newline rather than any last character in the line.  This fixes a potential bug for the last message in the file if the file doesn't end in a newline.