Match against /go argument when determining $Q
I've recently started using `go2.pl` (a drop-in replacement to `go.pl`),
which provides the handy `/go` command to switch windows. Consequently,
I can now do:

```
/bind meta-s-meta-i command go #irssi
```

to switch to the window containing the best match for item `#irssi`.

This patch teaches AWL to read the `/bind` output for such keybindings
as well, so that `$Q` works accordingly.

Signed-off-by: martin f. krafft <madduck@madduck.net>