Implement attention span
Rules can now be given an "attention span" such that they apply only
a certain period following your own engagement in a channel. For
instance, you might ask a question in a channel you'd usually only track
for hilights, and thus want to see activity for all messages for the
next hour.

Signed-off-by: martin f. krafft <madduck@madduck.net>