Rework command line parser
Provide a unified way to parse Irssi command lines for all our commands.
In theory, this means that arguments like rule position will be parsed
also for e.g. /ctrlact query, where this makes no sense, but so what. If
there's a problem, we can react to it in the command handler.

Signed-off-by: martin f. krafft <madduck@madduck.net>